Add dark theme powershell script
This script requires you to run Powershell using elevated privileges. If you run the script and Powershell is not being ran as administrator, the script will close the current Powershell script, re-open Powershell as admin and execute the script. The script comments out the line that disables the dark theme in the ssms.pkgundef file.